firestr now builds with VC 12
Ignore-this: be4e52b56677c09af4c55599bf8a4271

darcs-hash:20140217205412-c21fe-8b53b06aca47672d41959bf21acabbf9d346fd00

+++ src/gui/lua/widgets.cpp
@@ -582,7 +582,7 @@ namespace fire
             draw_view::draw_view(draw_ref ref, int width, int height) : 
                 QGraphicsView(), _ref{ref}, _button{0}
             {
-                setScene(new QGraphicsScene{0.0,0.0,width,height});
+                setScene(new QGraphicsScene{0.0,0.0,static_cast<qreal>(width),static_cast<qreal>(height)});
                 setMinimumSize(width,height);
                 setMouseTracking(true);
                 setRenderHint(QPainter::Antialiasing);
